Made search4 a module and fixed color on Windows
